Serving resources from server for offline
Leaflet and React resources are now served from the main stentor server
rather than on relying on a CDN.

While CDNs are useful for saving bandwidth upon mass adoption of an
application, it is better for new applications to ensure that fixed
stable versions are provided to prevent surprises during demos.

But mostly, this is to facilitate offline development.